Split meta column into two separate columns
