+++ frontend/src/pages/project/Snapshots/components/cells/DaysSince.vue
@@ -1,0 +1,22 @@
+<template>
+    <span>{{ since }}</span>
+</template>
+
+<script>
+import daysSince from '@/utils/daysSince'
+
+export default {
+    name: 'DaysSince',
+    props: {
+        date: {
+            required: true,
+            type: String
+        }
+    },
+    computed: {
+        since: function () {
+            return daysSince(this.date)
+        }
+    }
+}
+</script>
